Remove git log from CommandRunner's reponsibility
