+++ src/main/java/reposense/system/CommandRunner.java
@@ -1,43 +1,21 @@
 package reposense.system;
 
+import static reposense.util.StringsUtil.addQuote;
+
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.text.DateFormat;
-import java.text.SimpleDateFormat;
 import java.util.Date;
-import java.util.List;
 
 import reposense.git.CommitNotFoundException;
-import reposense.model.Author;
-import reposense.model.RepoConfiguration;
+import reposense.git.Util;
 import reposense.util.FileUtil;
 import reposense.util.StringsUtil;
 
 public class CommandRunner {
-    private static final DateFormat GIT_LOG_SINCE_DATE_FORMAT = new SimpleDateFormat("yyyy-MM-dd'T'00:00:00+08:00");
-    private static final DateFormat GIT_LOG_UNTIL_DATE_FORMAT = new SimpleDateFormat("yyyy-MM-dd'T'23:59:59+08:00");
-
-    // ignore check against email
-    private static final String AUTHOR_NAME_PATTERN = "^%s <.*>$";
-    private static final String OR_OPERATOR_PATTERN = "\\|";
 
     private static boolean isWindows = isWindows();
-
-    public static String gitLog(RepoConfiguration config, Author author) {
-        Path rootPath = Paths.get(config.getRepoRoot());
-
-        String command = "git log --no-merges -i ";
-        command += convertToGitDateRangeArgs(config.getSinceDate(), config.getUntilDate());
-        command += " --pretty=format:\"%H|%aN|%ad|%s\" --date=iso --shortstat";
-        command += convertToFilterAuthorArgs(author);
-        command += convertToGitFormatsArgs(config.getFormats());
-        command += convertToGitExcludeGlobArgs(author.getIgnoreGlobList());
-
-        return runCommand(rootPath, command);
-    }
-
     public static void checkout(String root, String hash) {
         Path rootPath = Paths.get(root);
         runCommand(rootPath, "git checkout " + hash);
@@ -56,7 +34,7 @@ public class CommandRunner {
         Path rootPath = Paths.get(root);
 
         String substituteCommand = "git rev-list -1 --before="
-                + GIT_LOG_UNTIL_DATE_FORMAT.format(untilDate) + " " + branchName;
+                + Util.GIT_LOG_UNTIL_DATE_FORMAT.format(untilDate) + " " + branchName;
         String hash = runCommand(rootPath, substituteCommand);
         if (hash.isEmpty()) {
             throw new CommitNotFoundException("Commit before until date is not found.");
@@ -101,7 +79,7 @@ public class CommandRunner {
 
         Path rootPath = Paths.get(root);
         String revListCommand = "git rev-list -1 --before="
-                + GIT_LOG_SINCE_DATE_FORMAT.format(date) + " " + branchName;
+                + Util.GIT_LOG_SINCE_DATE_FORMAT.format(date) + " " + branchName;
         return runCommand(rootPath, revListCommand);
     }
 
@@ -118,7 +96,7 @@ public class CommandRunner {
     public static String getShortlogSummary(String root, Date sinceDate, Date untilDate) {
         Path rootPath = Paths.get(root);
         String command = "git log --pretty=short";
-        command += convertToGitDateRangeArgs(sinceDate, untilDate);
+        command += Util.convertToGitDateRangeArgs(sinceDate, untilDate);
         command += " | git shortlog --summary";
 
         return runCommand(rootPath, command);
@@ -130,7 +108,7 @@ public class CommandRunner {
         return runCommand(rootPath, "git clone " + addQuote(location));
     }
 
-    private static String runCommand(Path path, String command) {
+    public static String runCommand(Path path, String command) {
         ProcessBuilder pb = null;
         if (isWindows) {
             pb = new ProcessBuilder()
@@ -170,72 +148,7 @@ public class CommandRunner {
         }
     }
 
-    private static String addQuote(String original) {
-        return "\"" + original + "\"";
-    }
-
     private static boolean isWindows() {
         return (System.getProperty("os.name").toLowerCase().indexOf("win") >= 0);
-    }
-
-    /**
-     * Returns the {@code String} command to specify the date range of commits to analyze for `git` commands.
-     */
-    private static String convertToGitDateRangeArgs(Date sinceDate, Date untilDate) {
-        String gitDateRangeArgs = "";
-
-        if (sinceDate != null) {
-            gitDateRangeArgs += " --since=" + addQuote(GIT_LOG_SINCE_DATE_FORMAT.format(sinceDate));
-        }
-        if (untilDate != null) {
-            gitDateRangeArgs += " --until=" + addQuote(GIT_LOG_UNTIL_DATE_FORMAT.format(untilDate));
-        }
-
-        return gitDateRangeArgs;
-    }
-
-    /**
-     * Returns the {@code String} command to specify the authors to analyze for `git log` command.
-     */
-    private static String convertToFilterAuthorArgs(Author author) {
-        StringBuilder filterAuthorArgsBuilder = new StringBuilder(" --author=\"");
-
-        // git author names may contain regex meta-characters, so we need to escape those
-        author.getAuthorAliases().stream()
-                .map(authorAlias -> String.format(AUTHOR_NAME_PATTERN,
-                        StringsUtil.replaceSpecialSymbols(authorAlias, ".")) + OR_OPERATOR_PATTERN)
-                .forEach(filterAuthorArgsBuilder::append);
-
-        filterAuthorArgsBuilder.append(
-                String.format(AUTHOR_NAME_PATTERN,
-                        StringsUtil.replaceSpecialSymbols(author.getGitId(), "."))).append("\"");
-        return filterAuthorArgsBuilder.toString();
-    }
-
-    /**
-     * Returns the {@code String} command to specify the file formats to analyze for `git` commands.
-     */
-    private static String convertToGitFormatsArgs(List<String> formats) {
-        StringBuilder gitFormatsArgsBuilder = new StringBuilder();
-        final String cmdFormat = " -- " + addQuote("*.%s");
-        formats.stream()
-                .map(format -> String.format(cmdFormat, format))
-                .forEach(gitFormatsArgsBuilder::append);
-
-        return gitFormatsArgsBuilder.toString();
-    }
-
-    /**
-     * Returns the {@code String} command to specify the globs to exclude for `git log` command.
-     */
-    private static String convertToGitExcludeGlobArgs(List<String> ignoreGlobList) {
-        StringBuilder gitExcludeGlobArgsBuilder = new StringBuilder();
-        final String cmdFormat = " " + addQuote(":(exclude)%s");
-        ignoreGlobList.stream()
-                .filter(item -> !item.isEmpty())
-                .map(ignoreGlob -> String.format(cmdFormat, ignoreGlob))
-                .forEach(gitExcludeGlobArgsBuilder::append);
-
-        return gitExcludeGlobArgsBuilder.toString();
     }
 }
