+++ src/test/java/reposense/system/CommandRunnerTest.java
@@ -4,7 +4,6 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Calendar;
-import java.util.Collections;
 import java.util.Date;
 
 import org.junit.Assert;
@@ -12,7 +11,6 @@ import org.junit.Test;
 
 import reposense.git.GitDownloader;
 import reposense.git.GitDownloaderException;
-import reposense.model.Author;
 import reposense.model.RepoConfiguration;
 import reposense.parser.ArgsParser;
 import reposense.parser.InvalidLocationException;
@@ -35,92 +33,6 @@ public class CommandRunnerTest extends GitTestTemplate {
         CommandRunner.checkout(config.getRepoRoot(), "test");
         Path branchFile = Paths.get(config.getRepoRoot(), "inTestBranch.java");
         Assert.assertTrue(Files.exists(branchFile));
-    }
-
-    @Test
-    public void gitLog_existingFormats_hasContent() {
-        String content = CommandRunner.gitLog(config, getAlphaAllAliasAuthor());
-        Assert.assertFalse(content.isEmpty());
-    }
-
-    @Test
-    public void gitLog_nonExistingFormats_noContent() {
-        config.setFormats(Collections.singletonList("py"));
-        String content = CommandRunner.gitLog(config, getAlphaAllAliasAuthor());
-        Assert.assertTrue(content.isEmpty());
-    }
-
-    @Test
-    public void gitLog_includeAllJavaFiles_success() {
-        config.setFormats(Collections.singletonList("java"));
-        String content = CommandRunner.gitLog(config, getAlphaAllAliasAuthor());
-        Assert.assertTrue(TestUtil.compareNumberExpectedCommitsToGitLogLines(8, content));
-    }
-
-    @Test
-    public void gitLog_fakeAuthorNameOnly_success() {
-        Author fakeAuthorName = new Author(FAKE_AUTHOR_NAME);
-
-        String content = CommandRunner.gitLog(config, fakeAuthorName);
-        Assert.assertTrue(TestUtil.compareNumberExpectedCommitsToGitLogLines(4, content));
-    }
-
-    @Test
-    public void gitLog_authorNameIncorrectCase_success() {
-        Author fakeAuthorName = new Author(FAKE_AUTHOR_NAME.toUpperCase());
-
-        String content = CommandRunner.gitLog(config, fakeAuthorName);
-        Assert.assertTrue(TestUtil.compareNumberExpectedCommitsToGitLogLines(4, content));
-    }
-
-    @Test
-    public void gitLog_fakeAuthorNameWithSpecialCharacter_success() {
-        Author fakeAuthorWithSpecialCharacter = new Author(FAKE_AUTHOR_NAME.replace("fake", "#()!"));
-
-        String content = CommandRunner.gitLog(config, fakeAuthorWithSpecialCharacter);
-        Assert.assertTrue(TestUtil.compareNumberExpectedCommitsToGitLogLines(4, content));
-    }
-
-    @Test
-    public void gitLog_includeAllJavaFilesAuthorIgnoreMovedFile_success() {
-        config.setFormats(Collections.singletonList("java"));
-        Author ignoreMovedFileAuthor = getAlphaAllAliasAuthor();
-        ignoreMovedFileAuthor.setIgnoreGlobList(Collections.singletonList("**movedFile.java"));
-
-        String content = CommandRunner.gitLog(config, ignoreMovedFileAuthor);
-        Assert.assertTrue(TestUtil.compareNumberExpectedCommitsToGitLogLines(6, content));
-    }
-
-    @Test
-    public void gitLog_authorIgnoreAllJavaFiles_success() {
-        Author ignoreAllJavaFilesAuthor = getAlphaAllAliasAuthor();
-        ignoreAllJavaFilesAuthor.setIgnoreGlobList(Collections.singletonList("*.java"));
-
-        String content = CommandRunner.gitLog(config, ignoreAllJavaFilesAuthor);
-        Assert.assertTrue(TestUtil.compareNumberExpectedCommitsToGitLogLines(2, content));
-    }
-
-    @Test
-    public void gitLog_authorWithAllCharactersRegexAlias_emptyResult() {
-        Author authorWithAllCharactersRegexAlias = new Author("none");
-        authorWithAllCharactersRegexAlias.setAuthorAliases(Collections.singletonList(".*"));
-
-        String content = CommandRunner.gitLog(config, authorWithAllCharactersRegexAlias);
-        Assert.assertTrue(content.isEmpty());
-    }
-
-    @Test
-    public void gitLog_sinceDateInFuture_noContent() {
-        Date date = TestUtil.getDate(2050, Calendar.JANUARY, 1);
-        config.setSinceDate(date);
-        String content = CommandRunner.gitLog(config, getAlphaAllAliasAuthor());
-        Assert.assertTrue(content.isEmpty());
-
-        date = TestUtil.getDate(1950, Calendar.JANUARY, 1);
-        config.setUntilDate(date);
-        config.setSinceDate(null);
-        content = CommandRunner.gitLog(config, getAlphaAllAliasAuthor());
-        Assert.assertTrue(content.isEmpty());
     }
 
     @Test
